CONTRIB-8674: Add test for recording upgrade routine (#470)
* Add tests for recording upgrade
* Minor fixes (groupid and return value)

+++ classes/task/upgrade_imported_recordings.php
@@ -17,6 +17,7 @@
 namespace mod_bigbluebuttonbn\task;
 
 use core\task\adhoc_task;
+use mod_bigbluebuttonbn\instance;
 use mod_bigbluebuttonbn\local\proxy\recording_proxy;
 
 /**
@@ -46,6 +47,7 @@ class upgrade_imported_recordings extends adhoc_task {
 
         // Magic number should be increased before release.
         $chunksize = 10;
+        $expected = $chunksize; // Expected number of records to process.
 
         // Initialize counter.
         $recordscount = 0;
@@ -58,47 +60,52 @@ class upgrade_imported_recordings extends adhoc_task {
                 LIMIT {$chunksize};";
         $logs = $DB->get_records_sql($sql);
 
-        // Get instanceids from the fetched logs.
-        $recsbyinstanceid = [];
-        foreach ($logs as $log) {
-            $recsbyinstanceid[$log->bigbluebuttonbnid] = (array)$log;
-        }
-        $instanceids = array_keys($recsbyinstanceid);
-
-        // Fetch instances that correspnd to instanceids selected.
-        $instances = $DB->get_records_list('bigbluebuttonbn', 'id', $instanceids);
-
-        // Create an instance of bigbluebuttonbn_recording per valid imported recording.
-        mtrace("Creating new recording records...");
-        foreach ($logs as $log) {
-            $recording = json_decode($log->meta, true)['recording'];
-            $newrecording = array(
-                'courseid' => $log->courseid,
-                'bigbluebuttonbnid' => $log->bigbluebuttonbnid,
-                'groupid' => 0, // The groupid should be taken from the meetingID.
-                'recordingid' => $recording['recordID'],
-                'status' => 2,
-                'imported' => 1,
-            );
-            if ($DB->count_records('bigbluebuttonbn_recordings', $newrecording) == 0) {
-                // Set headless flag to 1 if activity does not exist.
-                $newrecording['headless'] = !in_array($log->bigbluebuttonbnid, (array)$instances) ? 1 : 0;
-                $newrecording['importeddata'] = json_encode($recording);
-                $newrecording['timecreated'] = $log->timecreated;
-                $newrecording['timemodified'] = $log->timecreated;
-                $newrecording = $DB->insert_record('bigbluebuttonbn_recordings', $newrecording);
-                mtrace(json_encode($newrecording));
+        if ($logs) {
+            // Get instanceids from the fetched logs.
+            $recsbyinstanceid = [];
+            foreach ($logs as $log) {
+                $recsbyinstanceid[$log->bigbluebuttonbnid] = (array) $log;
             }
-        }
+            $instanceids = array_keys($recsbyinstanceid);
 
-        // Delete processed logs.
-        mtrace("Deleting migrated log records...");
-        foreach ($logs as $log) {
-            mtrace(json_encode($DB->get_records('bigbluebuttonbn_logs', array('id' => $log->id))));
-            $DB->delete_records('bigbluebuttonbn_logs', array('id' => $log->id));
-            $recordscount++;
-        }
+            // Fetch instances that correspnd to instanceids selected.
+            $instances = $DB->get_records_list('bigbluebuttonbn', 'id', $instanceids);
 
-        return ($recordscount == $chunksize);
+            // Create an instance of bigbluebuttonbn_recording per valid imported recording.
+            mtrace("Creating new recording records...");
+            foreach ($logs as $log) {
+                $recording = json_decode($log->meta, true)['recording'];
+                $instance = instance::get_from_meetingid($recording['meetingID']);
+                $newrecording = array(
+                    'courseid' => $instance->get_course_id(),
+                    'bigbluebuttonbnid' => $instance->get_instance_id(),
+                    'groupid' => $instance->get_group_id(), // The groupid should be taken from the meetingID.
+                    'recordingid' => $recording['recordID'],
+                    'status' => 2,
+                    'imported' => 1,
+                );
+                if ($DB->count_records('bigbluebuttonbn_recordings', $newrecording) == 0) {
+                    // Set headless flag to 1 if activity does not exist.
+                    $newrecording['headless'] = !in_array($log->bigbluebuttonbnid, (array) $instances) ? 1 : 0;
+                    $newrecording['importeddata'] = json_encode($recording);
+                    $newrecording['timecreated'] = $log->timecreated;
+                    $newrecording['timemodified'] = $log->timecreated;
+                    $newrecording = $DB->insert_record('bigbluebuttonbn_recordings', $newrecording);
+                    mtrace(json_encode($newrecording));
+                }
+            }
+
+            // Delete processed logs.
+            mtrace("Deleting migrated log records...");
+            foreach ($logs as $log) {
+                mtrace(json_encode($DB->get_records('bigbluebuttonbn_logs', array('id' => $log->id))));
+                $DB->delete_records('bigbluebuttonbn_logs', array('id' => $log->id));
+                $recordscount++;
+            }
+            $expected = count($logs) < $chunksize ? count($logs) : $chunksize;
+        } else {
+            $expected = 0;
+        }
+        return ($recordscount == $expected);
     }
 }
